cleanup Accounts.java
