+++ Core/build.xml
@@ -26,7 +26,6 @@
         <copy file="${env.TSK_HOME}/bindings/java/lib/sqlite-jdbc-3.8.11.jar" tofile="${basedir}/release/modules/ext/sqlite-jdbc-3.8.11.jar"/>
     </target>
 
-
     <target name="download-binlist">
         <get src="https://raw.githubusercontent.com/binlist/data/master/ranges.csv" 
              dest="src\org\sleuthkit\autopsy\datamodel"
